% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_bar_plot.R
\name{class_bar_plot}
\alias{class_bar_plot}
\title{Metabolic classification statistics}
\usage{
class_bar_plot(classlist, fontsize = 20, labelsize = 7)
}
\arguments{
\item{classlist}{a vector to class}

\item{fontsize}{font size}

\item{labelsize}{label size}
}
\value{
b
}
\description{
Metabolic classification statistics
}
\details{
Title
}
\examples{
a <- rep(LETTERS[1:10],sample(1:10, 10, replace = TRUE))
class_bar_plot(a)
class_bar_plot(classlist = mtcars$cyl,fontsize = 20,labelsize = 7)
}
