% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_rowname.R
\name{col_to_rowname}
\alias{col_to_rowname}
\title{Row name of a col}
\usage{
col_to_rowname(data, col)
}
\arguments{
\item{data}{Numerical matrix}

\item{col}{col}
}
\value{
data
}
\description{
Row name of a col
}
\details{
Title
}
\examples{
mtcars <- data.frame(rownames(mtcars),mtcars)
col_to_rowname(mtcars,1)
}
