% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Draw point line graph}
\usage{
line_plot(x, y, xlab = "x", ylab = "y")
}
\arguments{
\item{x}{a numerical vector.}

\item{y}{a numerical vector.}

\item{xlab}{x axis title}

\item{ylab}{y axis title}
}
\value{
data
}
\description{
Draw point line graph
}
\details{
Title
}
\examples{
line_plot(1:10,1:10)
}
