% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabonomics_analysis.R
\name{metabonomics_analysis}
\alias{metabonomics_analysis}
\title{Metabonomics analysis}
\usage{
metabonomics_analysis(a, group)
}
\arguments{
\item{a}{a matrix}

\item{group}{group}
}
\value{
b
}
\description{
Metabonomics analysis
}
\examples{
metabonomics_analysis(a,group)
}
