% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_cor_plot.R
\name{one_to_rowname}
\alias{one_to_rowname}
\title{Correlation analysis}
\usage{
one_to_rowname(data, col)
}
\arguments{
\item{data}{Numerical matrix}

\item{col}{col}
}
\value{
data
}
\description{
Correlation analysis
}
\examples{
qc_cor_plot(mtcars)
}
